+++ source/extensions/isaacsim.core.experimental.objects/python/impl/shapes/plane.py
@@ -54,7 +54,9 @@ class Plane(Shape):
             If the input shape is smaller than expected, data will be broadcasted (following NumPy broadcast rules).
         axes: Axes (plane's axis along which the surface is aligned) (shape ``(N,)``).
             If the input shape is smaller than expected, data will be broadcasted (following NumPy broadcast rules).
-        colors: Display colors (shape ``(N, 3)``).
+        colors: Normalized RGB display colors (shape ``(N, 3)``) or case-insensitive string representations.
+            Supported string representations include hex codes and X11/CSS4 color names without spaces,
+            as well as any other format supported by Matplotlib. Alpha channel is ignored for string representations.
             If the input shape is smaller than expected, data will be broadcasted (following NumPy broadcast rules).
         positions: Positions in the world frame (shape ``(N, 3)``).
             If the input shape is smaller than expected, data will be broadcasted (following NumPy broadcast rules).
@@ -69,6 +71,7 @@ class Plane(Shape):
 
     Raises:
         ValueError: If resulting paths are mixed (existing and non-existing prims) or invalid.
+        ValueError: Invalid string representation format for the colors.
         AssertionError: If wrapped prims are not USD Plane.
         AssertionError: If both positions and translations are specified.
 
@@ -79,9 +82,9 @@ class Plane(Shape):
         >>> from isaacsim.core.experimental.objects import Plane
         >>>
         >>> # given an empty USD stage with the /World Xform prim,
-        >>> # create cyan planes at paths: /World/prim_0, /World/prim_1, and /World/prim_2
+        >>> # create dark cyan planes at paths: /World/prim_0, /World/prim_1, and /World/prim_2
         >>> paths = ["/World/prim_0", "/World/prim_1", "/World/prim_2"]
-        >>> prims = Plane(paths, colors=[0.0, 1.0, 1.0])  # doctest: +NO_CHECK
+        >>> prims = Plane(paths, colors="DarkCyan")  # doctest: +NO_CHECK
     """
 
     def __init__(
@@ -93,7 +96,7 @@ class Plane(Shape):
         lengths: float | list | np.ndarray | wp.array | None = None,
         axes: Literal["X", "Y", "Z"] | list[Literal["X", "Y", "Z"]] | None = None,
         # Shape
-        colors: list | np.ndarray | wp.array | None = None,
+        colors: str | list | np.ndarray | wp.array | None = None,
         # XformPrim
         positions: list | np.ndarray | wp.array | None = None,
         translations: list | np.ndarray | wp.array | None = None,