+++ source/extensions/isaacsim.core.experimental.objects/python/tests/test_shape.py
@@ -13,10 +13,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from typing import Literal
+
 import isaacsim.core.experimental.utils.stage as stage_utils
 import omni.kit.commands
 import omni.kit.test
 from isaacsim.core.experimental.objects import Capsule, Cone, Cube, Cylinder, Plane, Shape, Sphere
+from isaacsim.core.experimental.prims.tests.common import cprint, draw_choice, draw_indices, parametrize
+
+
+async def populate_stage(max_num_prims: int, operation: Literal["wrap", "create"], **kwargs) -> None:
+    # create new stage
+    await stage_utils.create_new_stage_async()
+    # define prims
+    if operation == "wrap":
+        for i in range(max_num_prims):
+            stage_utils.define_prim(f"/World/A_{i}", "Sphere")
 
 
 class TestShape(omni.kit.test.AsyncTestCase):
@@ -60,3 +72,22 @@ class TestShape(omni.kit.test.AsyncTestCase):
         self.assertIsInstance(instances[4], Cylinder)
         self.assertIsInstance(instances[5], Plane)
         self.assertIsInstance(instances[6], Sphere)
+
+    @parametrize(backends=["usd"], prim_class=Sphere, populate_stage_func=populate_stage)
+    async def test_display_colors(self, prim, num_prims, device, backend):
+        choices = [
+            (0.1, 0.2, 0.3),  # RGB tuple
+            "#aBc",  # case-insensitive short hex RGB
+            "#0A1b2C",  # case-insensitive hex RGB
+            "0.5",  # grayscale
+            "k",  # basic color
+            "AquaMarine",  # case-insensitive X11/CSS4 color with no spaces
+            "xkcd:eggShell",  # case-insensitive  xkcd color
+            "tab:Green",  # case-insensitive tableau color
+            "C2",  # CN color specification
+            "none",  # special value (fully transparent)
+        ]
+        for indices, expected_count in draw_indices(count=num_prims, step=2):
+            cprint(f"  |    |-- indices: {type(indices).__name__}, expected_count: {expected_count}")
+            for v0, expected_v0 in draw_choice(shape=(expected_count,), choices=choices):
+                prim.set_display_colors(v0, indices=indices)